requirements.txt
autopep8==1.6.0
Django==4.0.1
django-crum==0.7.9
django-widget-tweaks==1.4.12
Pillow==9.0.0
weasyprint==52.5
whitenoise==6.2.0
gunicorn==20.1.0 
